Adding ER-diagram, Drucken

+++ db/migrate/20170615113528_create_printouts.rb
@@ -2,7 +2,7 @@ class CreatePrintouts < ActiveRecord::Migration[5.1]
   def change
     create_table :printouts do |t|
       t.references :report, foreign_key: true
-      t.int :times
+      t.bigint :times
 
       t.timestamps
     end
